refactor(scores): move buildSelectColumns tests to server/unit/, drop DB bootstrap
Moves the six pure buildSelectColumns unit tests from the server project
(scores-api-v3.servertest.ts) into a new server-unit file that skips the
Postgres globalSetup hook, following the polymorphicValue.servertest.ts
precedent and web/AGENTS.md guidance.
